requirements.txt
fastapi~=0.115.0
uvicorn~=0.32.0
sqlalchemy~=2.0.36
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
pydantic[email]~=2.10.0
httpx~=0.27.0
cachetools~=5.5.0
